Fix issue when refreshing a page
When refreshing a page in Firefox (not a hard refresh) the input value stayed the same. The issue was fastLiveFilter outputted the same as for an empty input. The callback function was called but the result list was unchanged / unfiltered

The fix here removes the callback(len); line and adds a triggerHandler to fully initialise. This worked for me. When the page is refreshed, it checks the current value of the input and does its thing.